requirements.txt
anthropic
bs4
python-dotenv
ftfy
flask
flask_socketio
Pillow
icecream
openai
pandas
playwright
pyautogui
pydantic
pygments
replicate
requests
rich
serpapi
tenacity